requirements.txt
absl-py==2.1.0
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.4.0
bleach==6.3.0
Bottleneck==1.4.2
brotlicffi==1.2.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.4.1
debugpy==1.8.16
decorator==5.2.1
dill==0.4.0
executing==2.2.1
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
grpcio==1.74.1
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.34.4
idna==3.11
iniconfig==2.3.0
ipykernel==6.31.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.6.3
jupyter_core==5.9.1
kaggle==2.0.0
kagglesdk==0.1.16
kiwisolver==1.4.9
lightning-utilities==0.15.3
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.2.1
mkl_fft==2.1.1
mkl_random==1.3.0
mkl-service==2.5.2
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.6.1
numexpr==2.14.1
numpy==2.3.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.6.77
onnx==1.20.1
onnx-ir==0.2.0
onnxruntime==1.24.3
onnxscript==0.6.2
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
pip==25.3
platformdirs==4.5.0
pluggy==1.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycares==4.10.0
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
PyQt6==6.9.1
PyQt6_sip==13.10.2
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytorch-lightning==2.6.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
sip==6.12.0
six==1.17.0
sniffio==1.3.1
stack_data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard_data_server==0.7.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.21.0
torch==2.9.1+cu126
torchmetrics==1.9.0
torchvision==0.24.1+cu126
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.5.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
uv==0.9.18
uvloop==0.22.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.5.0
yarl==1.22.0

